Flyttar første ting over
