+++ libs/etterlatte-brev-model/src/main/kotlin/no/nav/etterlatte/brev/Brev.kt
@@ -1,0 +1,19 @@
+package no.nav.etterlatte.brev.model
+
+typealias BrevID = Long
+
+enum class Status {
+    OPPRETTET,
+    OPPDATERT,
+    FERDIGSTILT,
+    JOURNALFOERT,
+    DISTRIBUERT,
+    SLETTET,
+}
+
+enum class BrevProsessType {
+    MANUELL,
+    REDIGERBAR,
+    AUTOMATISK,
+    OPPLASTET_PDF,
+}
